랜드마크 작성 관련

+++ src/main/java/edu/kh/bangbanggokgok/vo/board/LandMarkDetail.java
@@ -10,8 +10,8 @@ import lombok.Setter;
 public class LandMarkDetail {
 	
 	private int landMakrNo;
-	private int landMarkX;
-	private int landMarkY;
+	private double landMarkX;
+	private double landMarkY;
 	private String landMarkName;
 	private String landMarkContent;
 	private String landMarkAddress;
